Replace remaining avatars with robot fallback images
